fix(usage-metrics): batch over larger, configurable window
Previously flushing at `MAX_DELAY` of 2s was creating a much larger
volume of metrics than anticipated. This moves us to a default of a 20s
window over which we will aggregate in memory. This is configurable
using the `USAGE_METRICS_FLUSH_INTERVAL_SECS` environment variable.

With this change we will (still) aggregate in memory at most 1 entry
per component.

Additionally, buffering by a maximum number of events has been removed
entirely. This will further reduce the number of records we are
generating under load.

Ref: LOG-14685