fix(tokio): implement tokio task monitor workaround
There's an open issue with tokio where one task in the runtime can
prevent idle tasks from being scheduled on parked threads. Many of the
symptoms are similar to those that we see with the HTTP config reload
interval tick just not firing anymore. This change implements the work
around mentioned in the tokio issue.

Ref: LOG-14729