fix(usage): Use non-blocking mutex
The DB flusher is created using `tokio::spawn`, which provides execution
from the shared thread pool. This means that nothing should block the
threads or else it can block the reactor. The `DbFlusher` was using a
standard library `Mutex::lock()` that blocks until the lock is obtained.
Swapped it out for an async friendly tokio `Mutex`.

Ref: LOG-13733