fix(kafka): use a forked version of rdkafka
This is stopgap fix to address our needs in the short-term. I'm also
working on a better long-term fix that I hope we can get upstream so we
don't have to maintain a fork.

This uses a version of rdkafka that avoids an issue with blocking in
`rd_kafka_consumer_close()`. There's some issue in that call with the
way `rdkafka` is using it that causes it to indefinitely hang. In
addition to this I've created GitHub issues on both `librdkafka` and
`rust-rdkafka` to follow up.

The issues:

librdkafka: https://github.com/edenhill/librdkafka/issues/4041
rust-rdkafka: https://github.com/fede1024/rust-rdkafka/issues/509

Ref: LOG-14729