fix(gcp): rework flow for authentication error handling
For several sinks, including the GCP sinks, errors in the authentication
config are evaluated immediately at the time of configuration. This has
the effect of making it impossible to allow invalid credentials to be
entered by users by nature of how we are using Vector.

This moves this evaluation later in the flow such that a sink can be
"valid" with invalid credentials. Error messages are logged when the
sink is created, and the sink will not attempt to make (or retry) any
requests.

Ref: LOG-15040