ci: enable mezmo-specific integration tests
Upstream Vector maintains a test suite designed to run via Github
Actions. Early on we disabled them (by renaming to `*.upstream`) due
to the amount of time and resources the entire suite consumes.

This re-enables a subset of those tests in a workflow that will run on
the `master` branch, and can also be run manually. This is totally
separate from Jenkins, meaning that failing CI will not fail the build.
It does, however, give us an easy way to run the suite and keep tabs on
whether our changes have introduced bugs (or just broken a test) that is
only surfaced via runtime behavior.

Ref: LOG-15330