feat(transforms): Add `mezmo-reduce` transform
The out-of-the-box Vector functionality for reduce operates on the
root-level properties of the event. Since Mezmo only cares about what's
inside of .message, none of it worked. This commit copies reduce and
does the following things:

- Moves properties in and out of .message so that we can
utilize Vector's complex functionality for reducing events.

- Allows the user to specify which fields are dates and what
"template" (strftime) can be used to parse them.

- Since most date fields will be serialized as strings or integers,
automatically convert them into a Value::Timestamp during the reduce
process, but then serialize them back to the proper format on the way
out. This is mostly so we can take advantage of Vector's use of the
"_end" property which is created for each date at the end of the
reduce.

Ref: LOG-15465