feat(logging): Foundation for separate user logging stream
Creates a Mezmo specific context object that is part of the source,
transform and sink context objects that can be used to emit logging
events into a separate stream than the system logs. Includes a tower
service to wrap other services and log error responses to the Mezmo
specific stream if this is a user component.

Ref: LOG-15426