feat(vrl): added user_log function
Adds a user_log vrl function into the Mezmo section of vector and then
exposes that function in the remap transform. When used in the remap vrl
from a transform with a Mezmo pipeline component formatted id, the logs
will then produce into the user log stream. If used from a component
without a matching id, this turns into a no op function.

Ref: LOG-15435