fix(test): Fix user_log unit test
The test leveraged a macro from the vrl crate that are not available in
the application. Moving them into dev-dependencies broke the integration
tests, while moving them into the application seemed off. The solution
was to paste the test generated by the macro into the test function and
that required exposing a few trait impls.

Ref: LOG-15435