feat(sinks): Reformat LogEvent based on `message` property
This commit moves the functionality for "reshaping" an event from
the pipeline-service to the vector sinks. To hide the fact hat Mezmo
treats the `message` property as the event body, this contents of this
property will be moved to the root for selected sinks that have been
changed individually. Not all sinks have the same payload which is why
sinks have been individually changed. Those sinks NOT changed may have a
completely different payload structure, some of which require `message`.

Ref: LOG-15717