fix: Mezmo reshaping should only happen for JSON encoding
Previously, the message reshaping was not considering other codecs
like "text". Some of those rely on `message` as a special key, so we
should NOT reshape the payload. This changes the transformer method to
accept an optional Serializer to help decide what to do. This change
also makes Vector in line with *exactly* how the service used to do the
reshaping.

Ref: LOG-16032