fix: Adjust default value for bloom filter config
The tag cardinality limit transform was using a default value of 640kb,
for the bit vector, per key for the bloom filter. This is quite large.

This reduces it so that it has a false positive error rate of around 1%
for 500 values per key:

num_bits = 500 * ln(1/0.01) / (log(2)^2) = 4792.53 bits
num_bytes = num_bits/8 = ~600 bytes

Ref: LOG-15754