fix(s3): ensure batch status is updated
Previously, with acknowledgments enabled on a sink the S3 source would
get into a state where it was waiting for a message on a BatchNotifier
channel that would never be sent. That was because the notifier itself
was being cloned for each event in the batch, but the final reference
was not dropped prior to `await`ing the channel leading to a deadlock.

Ref: LOG-16241