feat(debug): additional debugging options for images
Configures the build process to use tokio_unstable and adds
tokio-console to the default feature set. This allows the inclusion of
tokio-console in the vector container image as an additional debugging
tool.

This change also installs gdb by default and uses a full debug build as
one more option for debugging production issues with vector.

Ref: LOG-16280