fix(mezmo_reduce): Correct flush logic and add size limitations
Reduce was not properly flushing its states. We discovered that the
setting for `expire_after_ms` relied on a gap in the data for it to
kick in and flush. Furthermore, that timer was always reset when new
data came in, creating a condition of never flushing as long as new
data was flowing. This commit removes that resetting and relies on
`expire_after_ms` to continually flush based on that interval.

Also, there were no memory bounds on how much reduce states can grow.
With some merge strategies like Array, the size can continually grow if
the `expire_after_ms` is large enough. This commit also adds byte size
estimates to each merge strategy so that we can enforce limits and
force flush if needed.

Ref: LOG-16319
Ref: LOG-16318