fix(mezmo_reduce): `group_by` fields should not have merge strategies
When using `group_by`, separate states are made to track the data by
those properties. However, the default merge strategies were still being
applied, so it causes unintended results. For string fields, this went
unnoticed because the default behavior for most types is
`DiscardMerger,` but if a field (like `status`) is an Integer, it was
using the `sum` default merger. This commit sets the merge strategy for
any `group_by` field to `DiscardMerger`, and ignores any custom merge
strategies for those fields.

Ref: LOG-16477