fix(mezmo_reduce): Translate field paths in `group_by` values
Since `group_by` is just a String vector, the logic to assign a Discard
merger to any group_by properties can fail if it contains field paths.
For example, if `status` is an integer, and we group by `.status`, the
code to assign the Discard merger will fail because `status !== .status`
(the properties in the incoming event obviously are not dot-notated).
This commit uses `parse_lookup` to translate field paths into raw
property names for that comparison. Since reduce only operates on the
root properties, we can ignore `parse_lookup` fields beyond the first
one (which will be at the root level of the event).

Ref: LOG-16614