feat(allocation-tracing): Re-enable alloc tracing
This re-enables the allocation tracing feature, but with support for a
much larger number of components, 4096 instead of 128. This is
accomplished by changing the allocation group ID to an `u16` type and
adjusting all the allocation tracking data structures. It's possible to
support more than 4096, but this seems to be a fair trade-off between
not being too big or too small, the full 2^16 range is almost certainly
too big.

Ref: LOG-16390