ci: add support for private `vrl` fork, rework Jenkins build
In order to support our VRL fork, we need to wire in credentials for the
various builds steps for unit tests, integration tests, and image
builds.

In addition, the new `vdev` CLI needed to be adjusted to use a newer
upstream base image, and force it to use `git` authorization for
`cargo`.

Ref: LOG-16178