fix: Add user log rate limiting
Adds rate limiting to user logs. Logs are rate limited by callsite and
not by the message. For this reason messages don't have to be static and
can use unique messages.

Adds a new CLI option (`--user-log-rate-limit`) and env var
(`VECTOR_USER_LOG_RATE_LIMIT`) for configuring the default rate limit
when not it's provided as an option to the `user_log()` function.

Ref: LOG-16826