fix(lib/codecs): cache metadata for prom_rw
Prometheus remote write clients do not include metric metadata in every
request, instead they periodically send an otherwise empty request with
the metadata describing the types of the metrics in the server. All
other requests contain type-erased "sample" types.

This PR adds a metadata cache to the deserializer so that it can work
out what the type of a given type-erased timeseries sample should be

Ref: LOG-17317