fix(mezmo_reduce): Cloned finalizers causing duplicate data
Whenever we are coercing events into new structures, we need to always
make sure to *not* clone the metadata. The ack finalizers are contained
in it, and cloning them can cause orphans which reject the batch and
cause data to be RE-read from Kafka. This commit makes sure to take the
finalizers when creating new log events.

Ref: LOG-16873