fix(mezmo_reduce): Allow field paths for merge strategies and dates
The default behavior for reduce is to only work with root-level
properties. This makes it cumbersome for users to enter properties that
do not allow path notation (since the rest of the system does). This
modifies an approach used to solve a similar problem for group_by and
creates a function to extract the root-level property when given a path
lookup.

This commit also includes an arithmetic fix for calculating sizes that
can become less over time (and thus have negative deltas). This problem
came up while writing a test for the field paths.

Ref: LOG-17480