fix(mezmo_reduce): Account for `message` schema prefix
Previously, we were not handling the valid VRL case which prepends the
"message" schema prefix. Something like `message.my_field` for either
`date_formats` or `group_by` was not working. Merge strategies were
already fixed in a previous commit.

This commit simply strips off the prefix for dates and group by fields,
since we move `message` contents in the root of a new LogEvent. The
prefix isn't needed at that point.

Ref: LOG-17505