fix(config): limit time vector waits for new topology to start
Add a configurable limit with a reasonable default value to the loop
that waits for a new topology to start. If the new topology can't start
in time, then the code will panic to kill the current process. The newly
started process should then be in a stable state to resume work as
normal. This is simply a work around to the config reload issues designed
to fix the issue without human involvement. We still need more work to
find the root cause.

Ref: LOG-17772