fix(config): use abort() instead of panic!() to kill vector
Initial testing and code review made it seem like the config reload path
was free of any handlers that would trap and suppress panic calls but that
wasn't the case in production. I noticed that one of the nodes hit the 150
second timeout but then just crashed the individual tokio thread instead
of crashing the process.

This commit uses the abort() function to cause an immediate termination
of the process, ignoring all the destructors.

Ref: LOG-17772