fix(shutdown): correct ShutdownSignal polling semantics
The ShutdownSignal poll() implementation contained an edge case where
the future mutates its internal state after the tripwire is resolved.
This new state will never ready a value and thus cause the ShutdownSignal
to act as if no signal was ever produced.

This commit detects and panics if this future is every polled after it
resolves a value. This is the only choice considering that the response
trigger would have been dropped on a previous poll() call. This is panic
is triggered inside the Exec source and is worked around by fusing the
shutdown signal before the Exec source loop polls on it.

Ref: LOG-17649