feat(prototype): Use global clustering ID
This change moves the `cluster_id` from a local calculation to a compact
hash of `template` to allow the `cluster_id` to be global across
multiple instances of the transform that are consuming the same stream
of logs. This is not perfect as the template can change as the logs are
streamed, but hopefully IDs will converge in most cases.

Ref: LOG-17981