feat(mezmo_config): VRL validation for remap tranforms
This adds a new configuration option, `validate_vrl`, to the Mezmo
config provider that enables validation of VRL in remap transforms. If a
bad VRL snippet is encountered then its pipeline is excluded from the
config topology. The user is also notified of a problem (if possible).

Includes two new metrics:
* `mezmo_config_vrl_validation_seconds` - a histogram that records
  VRL validation latency.
* `mezmo_config_vrl_validation_error` - a counter that tracks the number
  of VRL valdiation errors.

Ref: LOG-17690