chore(sinks): Mezmo sink add header for custom response
The ingesion service has a special (smarter) response for when the Mezmo
sink posts data. If all lines are errors, the overall HTTP code will
reflect that (400). This will cause the error to show up in user logs.

This change was tested locally with our Tilt setup, as we don't have the
ingestion service loaded here.

Ref: LOG-18256