feat(enrichment): Dynamic Pipelines can look up state variables
We have a need to let the user store arbitrary key/val pairs for each
pipeline. We will call them "state variables" and they can be used to
control the flow of a pipeline. This commit adds a new enrichment table
and a VRL function to wrap the `find_table_row` calls that pull data
from the enrichment table.

Because VRL is stateless (and not asynchronous), this implementation
will poll the postgres database directly, and load all state variables
for each account and pipeline into a Moka cache. There may be better
designs, but this is a POC, and it will yield 100% cache hits for the
VRL calls. Enrichment tables were chosen strictly because they offer an
async `build` context with which the DB can be connected.

Ref: LOG-18286
Ref: LOG-18333