fix(state_variables): Polling is killed on topology reload
When the topology is reloaded, it clones the `StateVariables` struct.
When the reload is complete, the newly-cloned struct is put into place
and the old one is dropped. Having the drop trait implemented was
stopping the polling flag, and since that data is behind an `Arc`
reference, the newly-cloned instance is still polling, but was exiting
early because the flag was shut off.

This removes the flag for stopping polling. We always need to be polling
or else the state data will be out of sync. Therefore, there's nothing
to do on drop.

Ref: LOG-18392