refactor(automated-parsing): use grok crate
This switches to use the `grok` crate, which matches what is used in VRL
for the `parse_grok` function. We will initially use the default
patterns baked into that library which are sufficient for this first
round.

We can use the `parse_grok` VRL function for parsing, which greatly
simplifies the implementation there without maintaining two lists of
patterns, and without needing to write another custom transform.
We already have precedent for this approach in our "Parse Sequentially"
processor.

This also avoids subtle differences that may exist between the
underlying regex libraries.

We are not precluded from adding our own patterns with this approach.

Ref: LOG-18677