fix(log_clustering): Always set the current node
In the case of wildcards, the current node was obtained but not
set. This attempts to finx the issue where it panic in dev with
test data. The specific panic could not be reproduced but a
bug that messed up with the structure was possible to be covered
in the test as part of this patch.

Ref: LOG-18799