fix(test): increased window size to ensure nothing expires
The test was failing in some cases on CI because the flushed result
wasn't empty. With only a 5 ms window, this could happen if the test
executed slowly. This changes the window size to 60 seconds to ensure
that execution would really have to slow down in order for the window
to expire.

Ref: LOG-18845