fix(s3 consolidation): large file copy source
summary: all unit tests pass for large files, but in dev
a 501 error is received stating that a header setting
implies functionality that isnt implemented. Research
points to this being either a permissions issue or that
the complete file range needs to be specified per the
current APIs. As such, the requester is now indicated
as acknowledged and the full file range is specified.
Also use the continuation token to handle >= 1000
file directories.

ref: LOG-18535