fix(otlp): replace or remove fields with an empty string
There is a bug during saving tap data as a sample
in a pipeline. The cause of this issue a 400 error
from the pipeline-service in case when payload contains
field/s with an empty string as a value. OTLP Logs, Metrics and Traces
could have empty strings in attributes/scope and other fields.
As a solution to fix this:
 1. Filter all the empty string fields out of `tags`
 2. Replace all the empty strings with `Null` for all other fields.

Ref: LOG-18908