fix(persistence): include pod name in db directory
In order for us to share a volume across all partitions (to start), each
partition needs a separate database. This will also facilitate moving
account across partitions in the future, since this process involves
pipelines for an account running in multiple partitions for a brief
period of time while buffers are drained and components are shut down.

Since we still have the HPA in play, and since the pod name includes the
partition name, the `POD_NAME` environment variable is used as the path
component.

Ref: LOG-19044