fix(persistence): define pvc per-partition
Previously the intent was to share a volume across all partitions to
start. This, however, has the consequence on scheduling of forcing
pods for all partitions to schedule on the same node.

Instead, this moves to a volume per-partition by defining a claim
template on the sts. In practice this will result in a volume per
pod (with a nominal replica count of 1) in each sts. Resources are now
also pulled from the partition definition in the `vector-partitions`
configmap.

Ref: LOG-19044