chore(tests): Re-enable splunk tests in Github Actions
A previous upgrade apparently "broke" the Splunk integration tests when
run in GHA. There have been several upstream merges since that time,
some of which made big changes to the compose files and `vdev` code for
running those tests. One such change was the addition of a matrix to test
multiple versions of Splunk. The tests appear to pass now, and without
having to specify version in the `env`. This commit simply turns on the
splunk tests in CI again.

Ref: LOG-18162