fix(http_server): Store request metadata in `PathPrefix::Metadata`
The original implementation for the `Legacy` namespace (which Mezmo
uses) dumps headers and query parameters into the root of the message.
Since Mezmo maintains a separate metadata field in th envelope, move those
values into the vector `metadata` for the Legacy namespace as it does
for Vector namespacing. This prevents the message from being polluted
and makes the data available to be inserted into the Mezmo envelope
downstream. The Edge http source will retrieve this metadata using the
`%` VRL operator when it constructs the `message` envelope (which does
not initially exist in the http_server).

Ref: LOG-19242