fix(ci): Speed improvements
Builds are taking too long. One reason is the "test image build" step
which can take 30 minutes. To hep dev flow, *only* do this during PRs so
that developers can just push branches to check CI/tests prior to
opening a PR that will do the test image build. Also, we regularly see
the '1 HOUR' timeouts and jobs are killed. This is a huge waste of time
and is usually due to CI servers not being up, but also cargo caches not
being primed. Increasing the timeout will allow more time for such
things.

Ref: LOG-19242