feat(classification): identify and annotate json string messages
For profiling, it was determined that a major gap we have is the
inability to pull annotations like `app` and `host` from the event
itself. We currently rely on the request to carry that information in
most cases. In other cases, the source (like splunk-hec) has a
known/fixed format that includes this information, and so we annotate
events during formatting off the queue.

Rather than going down the path of parsing all events/fields and
trying to map parsed data back into annotations, the decision was to
handle this for JSON only to start. From internal testing this will
support a large number of use-cases for profiling.

With this change, when the `message` is either an object or a JSON
string, we will look for app/host/level annotations in the object. The
fields we look for are configurable with some sane defaults.

Ref: LOG-19433