chore: disable integration tests on pull requests
We haven't found a ton of value in these tests running on every PR,
and they are somewhat expensive. Generally we should be running
integration tests locally when changing a component that has them.

This moves them to run on `master` only to provide at least some
visibility into failing tests.

Ref: LOG-19433