fix(classification): `SYSLOGLINE` is having false positive matches
The default pattern for `SYSLOGLINE` in the rust port of Grok is
very loosly-defined where most "matches" are just greedy data. This is
causing lots of lines from our agent to be matched as syslog even though
they are not. The rust port of Grok was originally based on a Java
library, but that library has a strictly-defined syslog pattern.

This commit adds the ability to override grok aliases with custom
definitions so that we can use the strict pattern from the Java library
as the definition for `SYSLOGLINE`.

Ref: LOG-19416