chore(ci): Publish to `docker.io` and use public repos
This commit prepares this repository to become the only place where
vector changes are done. The difference is that it will publish images
to `docker.io` where we can then use for Edge as well as our SaaS
product. This will require us to change our SaaS images to source from
dockerhub instead of gcr, but this is the first step towards that.

There were also changes to use public-only repos for dependencies. This
allows us to remove lots of github auth steps that are no longer needed.
This commit completely re-works `Jenkinsfile` to work from a public-only
repository standpoint.

Ref: LOG-18250