feat(mezmo): `user_log` acceps optional `captured_data` parameter
A parameter has been added to `user_log()` VRL (and macro) functions to
allow the user to specify some data that applies to the context of the
message. For example, an error in a `parse_json` VRL call may want to
provide the offending data that wasn't JSON.

The `captured_data` parameter is added for all log levels, although
primarily it will be most useful to include `captured_data` for errors
so that users can debug why a certain component is failing in their
pipeline.

Ref: LOG-19497