fix(classification): include metadata in event size calculation
Billing + profiling metrics track the user-accessible fields of the
event: `.message` and `.metadata`.

For classification we track the byte size of events analyzed during the
current run. Previously this only included the `.message` field, and
since we display the byte size from classification alongside the usage
metrics, these counts did not align.

Note that the byte size we track for log templates represents only the
line (string) itself, and as such it is expected to be different. The UI
will be updated to make this clearer.

Ref: LOG-19550