fix(ci): Clean disk space after ci runs [ci skip]
To avoid the disk filling up on the runner, call `volumes-clean`
at the end of each build. This is a new script geared toward clearing
the largest files accumulated in each build (vector-target). Since
the volumes are persisted for the lifetime of the agent, multiple builds
are writing to `vector-target` which causes large accumulation and the
inability to remove the volumes if another build is using it. Because of
this, `volumes-clean` is a new `Makefile` target that clears the volume,
which has been changed to respect the `VECTOR_TARGET` env value. We can
set this when running in CI so that each build has its own target
directory that can be cleared.

FWIW, the use of `cargo clean` isn't necessary because the version of `cargo`
on the host is different from the toolchain in the container. The
volumes for cargo-cache and rustup-cache can also be used multiple
times, therefore `make clean` is not appropriate to use.

Ref: LOG-19590