feat(throttle): reimplement throttle transform with basic rate limiting
The native `throttle` transform is backed by the `governor` library
which provides a very flexible GCRA implementation for rate limiting.
The library does not provide any way to explicitly initialize the quota
state, however, meaning that we cannot rehydrate the initial state of
the transform from state persistence (rocksdb).

For our purposes, we do not need the flexibity and advanced features
(e.g. burst rate capacity) provided by this crate. Instead, this adds
our own implementation `mezmo_throttle` which is identical, but instead
uses a very simple leaky bucket implementation. The state is represented
by a `HashMap` storing types that can natively be represented as a
JSON string for storage.

Ref: LOG-19565