fix(ci): Do no run tests on a release commit
Running tests on a release commit is too time consuming for this repo.
This commit adds a parent stage that can skip all tests with a `when`
clause, provided that the release commit title matches exactly what
semantic-release puts out, and that the commiter is our bot (so that
someone can't add a commit title that looks like a release commit, but
skips CI).

It's important to not the use of `git log` to get the top commit.
Previous version of this use the Jenkins `changelog` feature. However,
the value of that is ALL commits, not just the first one. When doing
upstream merges, this can include hundreds of commits that we don't want
subjected to this logic. Furthermore, `changelog` is no populated when a
branch/PR is first pushed which renders it useless.

Ref: LOG-19679