feat(tap): support VRL for tap event filtering
Currently we only support client-side filtering of tap events. This
makes it difficult to find events matching a pattern when those events
are infrequent and/or buried in a high-volume event stream.

This adds tap event filtering based on a VRL `Condition`, which can be
supplied to either the CLI or the GraphQL API. The service will generate
and supply this filter in the API requests based on user input.

```sh
vector tap --event-filter '.message.method == "POST"'
```

```graphql
outputEventsByComponentIdPatterns(
  outputsPatterns: ["remap"],
  filter: ".message.method == \"POST\""
) {
    ... on Log {
      type: __typename
      timestamp
      metadata: userMetadata
      message: json(field: ".message")
    }
}
```

Ref: LOG-19753