fix(http): handle and report partial failures with 207 status
The Mezmo destination uses the HTTP sink under the hood, and since the
`ingestion-service` handles partial failures by sending a 207 response,
we need to report errors from those requests.

This adds an explicit check in the `UserLoggingResponse` implementation
for the HTTP sink's `Request` type, which handles/extracts errors from
"successful" requests that contain a 207 response.

Ref: LOG-19799