fix(user_trace): Move the `captured_data` wrapper to the log fn
Storing captured data requires an `Object` wrapper keyed by
`"captured_data"` in order to preserve all data types in the DB.
Previously, this wrapper was added by the VRL function for `user_log`.
Therefore function calls within Vector to `user_log` required the
addition of that wrapper. This was an error, and the existence of that
wrapper should be a hidden implementation detail. This commit moves the
wrapper to the `MezmoUserLog` impl, which is used by function calls to
`user_log` as well as VRL calls to `user_log`.

Ref: LOG-19802