fix(postgresql): Nullify bad JSON serialization
There is a case where data can be serialized into a string, but after
the `.replace` of unicode nulls, it cannot be coerced back into JSON.
This is most likely bad `captured_data`, but that can't be confirmed.
Without a true use case to debug, assign `Null` to anything that fails
to go back into a `Serde::Value`.

Ref: LOG-19992