fix(otlp sink): make otlp sink healthcheck configurable
There is a limitation of the OTLP configuration
was found during testing of sending data
from OTLP destination to OTLP Source
within another pipeline. All the requests
failed because of healthcheck timeout error.
Our pipeline endpoint does not listen to any other
port except 80/443 but OTLP destination is configured
to send a healthcheck request to the 13133 port,
which cause an issue.
This PR changes the logic of the healthcheck,
now it sends a request to the same endpoint a customer
specified by default and gives an ability to configure
a specific port for the healthcheck.

Ref: LOG-20060