chore(otlp): get rid of the otlp healthcheck
The OTLP helthcheck has been enhanced recently to be able to send
requests to different ports depends on a sink config but recent testing
failed and to make the healthcheck work we have to do more enhancements
which will make it overcomplicated. Based on the discussion in Slack
(the link is specified in the ticket description) we decided to get rid
of the OTLP healthcheck.

Ref: LOG-20060