fix(kubernetes): pods rolling in order
Vector forms no cluster between them and thus does not benefit from the
ordered ready podManagementPolicy. On the contrary, it increases the
number of consumer rebalances that have to be done as the pods roll.

Instead, bring all the Vector pods up at once which should finish
rollouts faster with fewer rebalances.

Ref: INFRA-7234