feat(sources): `demo-logs` can work with user-provided data
A user should be able to supply their own data to be infinitely looped
in `demo-logs`. This can either be raw text, or serialized JSON (which
will ultimately get decoded as such).

Add a config option to allow `lines` to be supplied when `format` is
`user_provided`, and sequentially emit them in an infinite loop.

Ref: LOG-20639