fix(persistence): reuse rocksdb options instance
Previously, the RocksDB `Options` struct was recreated each time the db
handle was retreived from the registry. This causes the metrics to
"reset" when the component reloads, since the db handle now references a
new instance, and the original is dropped. The `Options` struct carries
the metrics data internally, and thus a single instance must be
maintained alongside each db instance.

Ref: LOG-20916