fix(mezmo::config): avoid walking the graph for input definitions
Previous the `input_definitions` function was called recursively for all
nodes in each pipeline graph. This function collects all schemas from
all ancestors of every node, then combines that collection to form
the "merged" schema of all ancestors. This was pulled from the topology
builder as part of the validation that occurs in our provider.

In `stage`, this has been observed to take up to 2 minutes and consumes
~100% of CPU while iterating over all components. This is with only 54
transforms present in the topology.

We are not leveraging component schemas for fields, and in practice all
of the schemas returned by `input_definitions` are the same anyway. This
change avoids collecting schemas an instead uses the static
`default_legacy_namespace` schema, which aligns with our usage.

Ref: LOG-21352