fix(resources): assign vector_threads from config
we should be limiting the `VECTOR_THREADS` environment variable,
available since the .33 release, to ensure that we do not have an
host-bounded vector thread issue.  we have run into issues where
vector instances located on very busy machines cause a lot of CPU
context switching, because of the aggregate threads asking for time
on the host machine.

incidents: #-incident-335521, #-incident-335205
ref: LOG-21403