fix(traces): maintain timestamp type
summary: child spans are serialized and kept until the
need to recall when the head span it received. Upon
deserialization, the timestamp was considered a string
instead of an actual DateTime which plagued components
which assumed its type. This corrects the timestamp of
each record before considering the event as output.

ref: LOG-21405