feat(aggregate_distributed): add initial implementation
This includes the initial implementation of the
`mezmo_aggregate_distributed` transform, which supports SUM and AVG
strategies for Metric-shaped log events. Gauge and Counter types
are supported initially. We will iterate on this to add new
strategies and support other metric types as-needed.

Both recording and flushing events to/from state are handled via Lua
scripts executed within the datastore.

Connections are handled via the `redis-rs#ConnectionManager` interface,
which supports automatic, transparent reconnection during failover or
server restarts. Retries are handled at the event-level to ensure each
event is recorded in state (up to a max timout) before moving on to the
next event in the stream, providing a form of "backpressure" in the
transform. Data-related errors are bubbled up to the user.

Internal metrics are also emitted to indicate the health and performance
of the transform w/r/t the datastore.

Ref: LOG-21473