feat(throttle_distributed): add mezmo_throttle_distributed transform
Adds the implementation of a sliding-window rate-limiter, operating
against Redis/Dragonflydb. This supports all of the same capabilities
as the existing `mezmo_throttle`, but with the rate-limit state
reflected out into the datastore.

Much like LOG-21473 retries are handled at both the connection and the
event level. This transform will "fail open" in the case where we have
extended connection interruption, letting all events to pass through
unthrottled. Maximum cardinality within the datastore is also enforced,
and when reached all new events/windows will be unthrottled.

Ref: LOG-21564