feat: Set interval_ms in to_metric function (#652)
* feat: Set interval_ms in to_metric function

Define a top level time field on a log event that
includes the interval_ms for rate type metrics.

Ref: LOG-21507

* refactor: Toggle Datadog series API via env var

Define an environment variable to control use of Datadog
series v2 API to egress metrics.
Update the config implementation to use v2 API only if
the env var is set to true.

Ref: LOG-21506