fix(aggregate_distributed): restructure output event value
Previously the `value` field contained the raw numeric value. Instead,
this field should be of the `Value::Object` variant, which carries the
`type` of the metric value in order to conform to the Metric model.
The `value_type` field is now included in the fields stored in
the window.

Additionally the root-level `timestamp` field is added to the output
events and contains the same value as the `window_end` timestamp.

Ref: LOG-21688