fix(rocksdb): set limits on log files
summary: the log files for state persistence grow unbounded
at the moment. Its unnecessary to keep all the info logs
for the database, so added configs to be used when establishing
the connection to limit the number and size of logs. Configs
are established at the partition level to allow per partition
customization. If omitted, defaults are established in the code.

ref: LOG-21391