fix(aggregate_distributed): adjust flush timing for delays
When aggregating events with delayed timestamps, the previous behavior
would cause new windows to be created and flushed almost immediately
near the flush boundary for the window. Even with a minor delay (vs the
wall-clock) this can cause multiple output events for the same window.

This introduces two changes to help mitigate this:

- The flush timestamp is now derived from `now`, rather than always
  calculated from `window_start_ts + window_duration_ms`. This ensures
  we always keep a window "open" for aggregation for _at least_ the
  window duration.
- A `flush_grace_period_ms` configuration option is added, which
  represents an additional delay before flushing an expired window.
  This provides some tolerance for our own processing delays, and gives
  us further control over the timing globally. Initial value is 30s.

Ref: LOG-21771